requirements.txt
strawberry-graphql
SQLAlchemy
psycopg[binary]
